refactor: reorganize scheduler and daemon with clean architecture and test coverage expansion
