+++ .gitignore
@@ -8,3 +8,4 @@ output
 main
 libwrapper.a
 gthulhu-cli
+monitor/bpf/vmlinux.h
