+++ internal/scheduler/policy/api_interval.go
@@ -1,0 +1,8 @@
+package policy
+
+func NormalizeAPIInterval(interval int, enabled bool) int {
+	if interval <= 0 || !enabled {
+		return 5
+	}
+	return interval
+}
